refactored last commit and deleted ipynb and refectored unwanted functions as well transformation using both pipeline working correctly
